# internal/oas/generated/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package gen

import (
	"fmt"
)

func (s *ErrorResponseStatusCode) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

type CreateTokensBadRequest ErrorResponse

func (*CreateTokensBadRequest) createTokensRes() {}

type CreateTokensNotFound ErrorResponse

func (*CreateTokensNotFound) createTokensRes() {}

// Ref: #/components/schemas/CreateTokensRequest
type CreateTokensRequest struct {
	AdditionalInfo CreateTokensRequestAdditionalInfo `json:"additionalInfo"`
}

// GetAdditionalInfo returns the value of AdditionalInfo.
func (s *CreateTokensRequest) GetAdditionalInfo() CreateTokensRequestAdditionalInfo {
	return s.AdditionalInfo
}

// SetAdditionalInfo sets the value of AdditionalInfo.
func (s *CreateTokensRequest) SetAdditionalInfo(val CreateTokensRequestAdditionalInfo) {
	s.AdditionalInfo = val
}

type CreateTokensRequestAdditionalInfo map[string]string

func (s *CreateTokensRequestAdditionalInfo) init() CreateTokensRequestAdditionalInfo {
	m := *s
	if m == nil {
		m = map[string]string{}
		*s = m
	}
	return m
}

// Ref: #/components/schemas/CreateTokensResponse
type CreateTokensResponse struct {
	// Jwt token.
	AccessToken string `json:"accessToken"`
	// Jwt token.
	RefreshToken string `json:"refreshToken"`
}

// GetAccessToken returns the value of AccessToken.
func (s *CreateTokensResponse) GetAccessToken() string {
	return s.AccessToken
}

// GetRefreshToken returns the value of RefreshToken.
func (s *CreateTokensResponse) GetRefreshToken() string {
	return s.RefreshToken
}

// SetAccessToken sets the value of AccessToken.
func (s *CreateTokensResponse) SetAccessToken(val string) {
	s.AccessToken = val
}

// SetRefreshToken sets the value of RefreshToken.
func (s *CreateTokensResponse) SetRefreshToken(val string) {
	s.RefreshToken = val
}

func (*CreateTokensResponse) createTokensRes()  {}
func (*CreateTokensResponse) refreshTokensRes() {}

// Ref: #/components/schemas/ErrorResponse
type ErrorResponse struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// GetCode returns the value of Code.
func (s *ErrorResponse) GetCode() int32 {
	return s.Code
}

// GetMessage returns the value of Message.
func (s *ErrorResponse) GetMessage() string {
	return s.Message
}

// SetCode sets the value of Code.
func (s *ErrorResponse) SetCode(val int32) {
	s.Code = val
}

// SetMessage sets the value of Message.
func (s *ErrorResponse) SetMessage(val string) {
	s.Message = val
}

// ErrorResponseStatusCode wraps ErrorResponse with StatusCode.
type ErrorResponseStatusCode struct {
	StatusCode int
	Response   ErrorResponse
}

// GetStatusCode returns the value of StatusCode.
func (s *ErrorResponseStatusCode) GetStatusCode() int {
	return s.StatusCode
}

// GetResponse returns the value of Response.
func (s *ErrorResponseStatusCode) GetResponse() ErrorResponse {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *ErrorResponseStatusCode) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetResponse sets the value of Response.
func (s *ErrorResponseStatusCode) SetResponse(val ErrorResponse) {
	s.Response = val
}

// NewOptCreateTokensRequest returns new OptCreateTokensRequest with value set to v.
func NewOptCreateTokensRequest(v CreateTokensRequest) OptCreateTokensRequest {
	return OptCreateTokensRequest{
		Value: v,
		Set:   true,
	}
}

// OptCreateTokensRequest is optional CreateTokensRequest.
type OptCreateTokensRequest struct {
	Value CreateTokensRequest
	Set   bool
}

// IsSet returns true if OptCreateTokensRequest was set.
func (o OptCreateTokensRequest) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptCreateTokensRequest) Reset() {
	var v CreateTokensRequest
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptCreateTokensRequest) SetTo(v CreateTokensRequest) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptCreateTokensRequest) Get() (v CreateTokensRequest, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptCreateTokensRequest) Or(d CreateTokensRequest) CreateTokensRequest {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

type RefreshTokensBadRequest ErrorResponse

func (*RefreshTokensBadRequest) refreshTokensRes() {}

type RefreshTokensForbidden ErrorResponse

func (*RefreshTokensForbidden) refreshTokensRes() {}

type RefreshTokensNotFound ErrorResponse

func (*RefreshTokensNotFound) refreshTokensRes() {}

type RefreshTokensReq struct {
	// JWT token.
	RefreshToken string `json:"refreshToken"`
	// JWT token.
	AccessToken string `json:"accessToken"`
}

// GetRefreshToken returns the value of RefreshToken.
func (s *RefreshTokensReq) GetRefreshToken() string {
	return s.RefreshToken
}

// GetAccessToken returns the value of AccessToken.
func (s *RefreshTokensReq) GetAccessToken() string {
	return s.AccessToken
}

// SetRefreshToken sets the value of RefreshToken.
func (s *RefreshTokensReq) SetRefreshToken(val string) {
	s.RefreshToken = val
}

// SetAccessToken sets the value of AccessToken.
func (s *RefreshTokensReq) SetAccessToken(val string) {
	s.AccessToken = val
}